clase 29 - insertar usando typeorm

+++ src/products/entities/product.entity.ts
@@ -6,7 +6,7 @@ export class Product {
 
     @Column('text', { unique: true }) title: string;
 
-    @Column('numeric', { default: 0 }) price: number;
+    @Column('float', { default: 0 }) price: number;
 
     @Column({ type: 'text', nullable: true }) description: string;
